lab2-adk-agent/pyproject.toml
[project]
name = "adk"
version = "0.1.0"
description = ""
authors = [
    {name = "ermadatov",email = "ermadatov@cloud.ru"}
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "google-adk (>=1.13.0,<2.0.0)",
    "a2a-sdk (>=0.3.4,<0.4.0)",
    "litellm (>=1.76.3,<2.0.0)",
    "openinference-instrumentation-google-adk (>=0.1.4,<0.2.0)",
    "arize-phoenix-otel (>=0.13.0,<0.14.0)"
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
google-adk = "^1.13.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
